feat: compensate device clock drift
Compensate the device clock drift by sending the current seconds since
epoch from the companion, which very likely has a more accurate system
time, to the device, where it's used to calculate the clock drift.

Many smart watches don't have a very precise real-time clock and
compensate for it by adjusting the device clock via network calls or
in the case of the Fitbit devices when they are synchronized with the
paired smartphone. Even if the device is synchronized regularly, time
differences within the multiple seconds range have been observed
during testing, which is problematic with tokens featuring a validity
time span of thirty seconds or less.